google_chrome_profiles: add path column
There is currently no column that can be used to link rows returned by the _google_chrome_profiles_ table with the rows returned by the OSQuery built-in _chrome_extensions_ table. This is useful for determining the email, and therefore the Google Account, associated with the profile, which can be used as a proxy for whether the profile is a work or non-work one.

This PR adds a _path_ column to the _google_chrome_profiles_ table. This is the path to the profile's directory, and should match _profile_path_ column in the _chrome_extensions_ table, allowing these two tables to be joined together. It does this by pulling out the profile's directory name from the same dictionary used to get the existing data, combines it with the Chrome user data directory path, and finally checks that it exists before returning it. It is therefore possible for this column to be empty, if for whatever reason, the computed path does not exist.

```
osquery> select * from google_chrome_profiles;
+----------+------------------------+------------------+-----------+-----------------------------------------------------------------+
| username | email                  | name             | ephemeral | path                                                            |
+----------+------------------------+------------------+-----------+-----------------------------------------------------------------+
| user1    | user1@test.example.com | test.example.com | 0         | /Users/user/Library/Application Support/Google/Chrome/Default   |
| user2    | user2@example.com      | example.com      | 0         | /Users/user/Library/Application Support/Google/Chrome/Profile 1 |
+----------+------------------------+------------------+-----------+-----------------------------------------------------------------+
```